Make Standalone Shell use picocli
