+++ tools/bootstrapper/build.gradle
@@ -11,7 +11,6 @@ dependencies {
     compile "info.picocli:picocli:$picocli_version"
     compile "org.slf4j:jul-to-slf4j:$slf4j_version"
     compile "org.apache.logging.log4j:log4j-slf4j-impl:$log4j_version"
-    compile "com.jcabi:jcabi-manifests:$jcabi_manifests_version"
 }
 
 jar {
